refactor: todos module
